add tests, fix bugs with Order entity, extend User entity

+++ src/main/java/ru/sfedu/sevenTravel/persistence/dao/genericDAO/GenericDAOImpl.java
@@ -1,24 +1,22 @@
 package ru.sfedu.sevenTravel.persistence.dao.genericDAO;
 
-import org.hibernate.Session;
-import org.hibernate.SessionFactory;
-import ru.sfedu.sevenTravel.utils.HibernateUtil;
+import ru.sfedu.sevenTravel.utils.JPAUtil;
 
 import javax.persistence.*;
 import javax.persistence.criteria.CriteriaQuery;
 import java.io.Serializable;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public abstract class GenericDAOImpl<T, ID extends Serializable> implements GenericDAO<T, ID> {
 
-    //public SessionFactory sessionFactory = HibernateUtil.getSessionFactory();
-    private EntityManager em;
+    protected EntityManager em;
 
     private final Class<T> entityClass;
 
     protected GenericDAOImpl(Class<T> entityClass){
         this.entityClass = entityClass;
-        this.em = HibernateUtil.getSessionFactory().createEntityManager();
+        this.em = JPAUtil.getSessionFactory().createEntityManager();
     }
 
     public void setEntityManager(EntityManager em){
@@ -59,18 +57,39 @@ public abstract class GenericDAOImpl<T, ID extends Serializable> implements Gene
 
     @Override
     public T makePersistent(T entity){
-
+        if (entity==null) throw new NullPointerException();
+        open();
         em.getTransaction().begin();
         entity = em.merge(entity);
         em.getTransaction().commit();
-
         return entity;
+    }
+
+    public List<T> makePersistent(List<T> entities){
+        if (entities==null) throw new NullPointerException();
+        if (entities.size()==0) throw new NullPointerException();
+        open();
+        em.getTransaction().begin();
+        entities = entities.stream().map(em::merge).collect(Collectors.toList());
+        em.getTransaction().commit();
+        return entities;
     }
 
     @Override
     public void makeTransient(T entity){
+        if (entity==null)throw new NullPointerException();
+        open();
         em.getTransaction().begin();
         em.remove(entity);
+        em.getTransaction().commit();
+    }
+
+    public void makeTransient(List<T> entities){
+        if (entities==null) throw new NullPointerException();
+        if (entities.size()==0) throw new NullPointerException();
+        open();
+        em.getTransaction().begin();
+        entities.forEach(em::remove);
         em.getTransaction().commit();
     }
 
@@ -84,9 +103,12 @@ public abstract class GenericDAOImpl<T, ID extends Serializable> implements Gene
         );
     }
     public void close(){
-        this.em.close();
+        if (em!=null && em.isOpen())
+            this.em.close();
     }
+
     public void open(){
-        em = HibernateUtil.getSessionFactory().createEntityManager();
+        if (em!=null && em.isOpen()) return;
+        em = JPAUtil.getSessionFactory().createEntityManager();
     }
 }
